Fix links to DEB and RPM packages
The links to the DEB and RPM packages are broken as they use standard
variables pointing to the latest IC version. DEB and RPM packages usually
run behind the latest IC releass.

Additional variables have been added to variable.txt which are static and
pointing to the latest DEB/RPM locations.

Also downloads.html has been updated to note the latest versions of the
DEB and RPM packages.