First version of script to check validity of sites in Hall of Fame
- Script currently can take commandline URL or read from a file (HF).
- Script will check existence of cookie containing MV_SESSION_ID on
  URL
- If no hit, script will check existence of cooking containing
  MV_SESSION_ID on URL/admin/index
- If no hit, site is considered non-interchange

Script is by no means finished. Possible future additions:
- Query icdev database directly instead of use of file
- Query icdev for URL and Link Test .. Sometimes curl will follow a
  301, but there are also META redirects etc.
- Other changes depeding on what future results will give with this
  setup.

It still leaves some manual work, but if a site is marked as NOT
Interchange it will need a visual check on the site and on the
description in case it is an Intranet solution where it does not
have an actual URL to connect to. For this the 'Link Test' can be
used and marked with 'none'.